[add] ball_laucher_plugin header
The reason why this plugin was not loaded is gazebo namespace was missed
for the class.
